Rename component to internal_temperature
