+++ esphome/components/internal_temperature/__init__.py
