+++ esphome/components/internal_temperature/sensor.py
@@ -8,14 +8,14 @@ from esphome.const import (
     ENTITY_CATEGORY_DIAGNOSTIC,
 )
 
-cpu_temperature_ns = cg.esphome_ns.namespace("cpu_temperature")
-CPUTemperatureSensor = cpu_temperature_ns.class_(
-    "CPUTemperatureSensor", sensor.Sensor, cg.PollingComponent
+internal_temperature_ns = cg.esphome_ns.namespace("internal_temperature")
+InternalTemperatureSensor = internal_temperature_ns.class_(
+    "InternalTemperatureSensor", sensor.Sensor, cg.PollingComponent
 )
 
 CONFIG_SCHEMA = cv.All(
     sensor.sensor_schema(
-        CPUTemperatureSensor,
+        InternalTemperatureSensor,
         unit_of_measurement=UNIT_CELSIUS,
         accuracy_decimals=1,
         device_class=DEVICE_CLASS_TEMPERATURE,
