+++ tests/test5.yaml
@@ -354,8 +354,8 @@ select:
       "Three": 3
 
 sensor:
-  - platform: cpu_temperature
-    name: "CPU Temperature"
+  - platform: internal_temperature
+    name: "Internal Temperature"
   - platform: selec_meter
     total_active_energy:
       name: SelecEM2M Total Active Energy
